+++ pkg/metadata/server/object.go
@@ -2,11 +2,7 @@ package server
 
 import (
 	"context"
-	"reflect"
 
-	yaml "gopkg.in/yaml.v2"
-
-	apitypes "github.com/runmachine-io/runmachine/pkg/api/types"
 	"github.com/runmachine-io/runmachine/pkg/errors"
 	"github.com/runmachine-io/runmachine/pkg/metadata/conditions"
 	pb "github.com/runmachine-io/runmachine/pkg/metadata/proto"
@@ -121,248 +117,6 @@ func (s *Server) ObjectList(
 	return nil
 }
 
-// validateObjectSetRequest ensures that the data the user sent in the
-// request's payload can be unmarshal'd properly into YAML, contains all
-// relevant fields.  and meets things like property meta validation checks.
-//
-// Returns a fully validated Object protobuffer message that is ready to send
-// to backend storage.
-func (s *Server) validateObjectSetRequest(
-	req *pb.ObjectSetRequest,
-) (*types.ObjectWithReferences, error) {
-	// reads the supplied buffer which contains a YAML document describing the
-	// object to create or update, and returns a pointer to an Object
-	// protobuffer message containing the fields to set on the new (or changed)
-	// object.
-	obj := &apitypes.Object{}
-	if err := yaml.Unmarshal(req.Payload, obj); err != nil {
-		return nil, err
-	}
-
-	// Simple input data validations
-	if obj.ObjectType == "" {
-		return nil, ErrObjectTypeRequired
-	}
-	if obj.Partition == "" {
-		return nil, ErrPartitionRequired
-	}
-
-	// Validate the referred to type, partition and project actually exist
-	part, err := s.store.PartitionGet(
-		// Look up by UUID *or* name...
-		&pb.PartitionFilter{
-			UuidFilter: &pb.UuidFilter{
-				Uuid: obj.Partition,
-			},
-			NameFilter: &pb.NameFilter{
-				Name: obj.Partition,
-			},
-		},
-	)
-	if err != nil {
-		if err == errors.ErrNotFound {
-			return nil, errPartitionNotFound(obj.Partition)
-		}
-		// We don't want to leak internal implementation errors...
-		s.log.ERR("failed when validating partition in object set: %s", err)
-		return nil, errors.ErrUnknown
-	}
-
-	objType, err := s.store.ObjectTypeGet(obj.ObjectType)
-	if err != nil {
-		if err == errors.ErrNotFound {
-			return nil, errObjectTypeNotFound(obj.ObjectType)
-		}
-		// We don't want to leak internal implementation errors...
-		s.log.ERR("failed when validating object type in object set: %s", err)
-		return nil, errors.ErrUnknown
-	}
-
-	objProperties := make([]*pb.Property, 0)
-	if obj.Properties != nil {
-		// Validate that the properties set against this object meet any schema
-		// associated with that property key and object type
-		for key, value := range obj.Properties {
-			prop, err := s.validateObjectProperty(part, objType, key, value)
-			if err != nil {
-				return nil, err
-			}
-			objProperties = append(objProperties, prop)
-		}
-	}
-
-	return &types.ObjectWithReferences{
-		Partition:  part,
-		ObjectType: objType,
-		Object: &pb.Object{
-			Partition:  part.Uuid,
-			ObjectType: objType.Code,
-			Project:    obj.Project,
-			Name:       obj.Name,
-			Uuid:       obj.Uuid,
-			Tags:       obj.Tags,
-			Properties: objProperties,
-		},
-	}, nil
-}
-
-// validateObjectProperty ensures that the supplied key and value meet any
-// defined property meta constraints that may have been defined for that
-// object type and key. Returns a pointer to a Property protobuffer message.
-func (s *Server) validateObjectProperty(
-	partition *pb.Partition,
-	objType *pb.ObjectType,
-	key string,
-	value string,
-) (*pb.Property, error) {
-	pds, err := s.store.PropertyDefinitionList(
-		[]*conditions.PropertyDefinitionCondition{
-			&conditions.PropertyDefinitionCondition{
-				PartitionCondition:   conditions.PartitionEqual(partition),
-				ObjectTypeCondition:  conditions.ObjectTypeEqual(objType),
-				PropertyKeyCondition: conditions.PropertyKeyEqual(key),
-			},
-		},
-	)
-	if err != nil {
-		return nil, err
-	}
-	if len(pds) > 0 {
-		pd := pds[0]
-		err := s.validateValueWithSchema(value, pd.Schema)
-		if err != nil {
-			return nil, errors.ErrFailedPropertyDefinitionValidation(key, err)
-		}
-	}
-	return &pb.Property{
-		Key:   key,
-		Value: value,
-	}, nil
-}
-
-// validateValueWithSchema returns an error if the supplied value passes the
-// supplied property meta document, nil otherwise.
-func (s *Server) validateValueWithSchema(
-	value string,
-	schema *pb.PropertySchema,
-) error {
-	return nil
-}
-
-func objectChanged(a, b *types.ObjectWithReferences) bool {
-	if a.Partition.Uuid != b.Partition.Uuid {
-		return true
-	}
-	if a.ObjectType.Code != b.ObjectType.Code {
-		return true
-	}
-	if a.Object.Uuid != b.Object.Uuid {
-		return true
-	}
-	if a.Object.Name != b.Object.Name {
-		return true
-	}
-	if a.Object.Project != b.Object.Project {
-		return true
-	}
-	if !reflect.DeepEqual(a.Object.Tags, b.Object.Tags) {
-		return true
-	}
-	if !reflect.DeepEqual(a.Object.Properties, b.Object.Properties) {
-		return true
-	}
-	return false
-}
-
-func (s *Server) ObjectSet(
-	ctx context.Context,
-	req *pb.ObjectSetRequest,
-) (*pb.ObjectSetResponse, error) {
-	// TODO(jaypipes): AUTHZ check if user can write objects
-
-	input, err := s.validateObjectSetRequest(req)
-	if err != nil {
-		return nil, err
-	}
-
-	// Determine if this is a new object the user is creating or an existing
-	// one they are updating
-	var existing *types.ObjectWithReferences
-	var cond *conditions.ObjectCondition
-	if input.Object.Uuid != "" {
-		cond = &conditions.ObjectCondition{
-			UuidCondition: conditions.UuidEqual(input.Object.Uuid),
-		}
-	} else {
-		cond = &conditions.ObjectCondition{
-			ObjectTypeCondition: conditions.ObjectTypeEqual(input.ObjectType),
-			NameCondition:       conditions.NameEqual(input.Object.Name),
-		}
-		if input.ObjectType.Scope == pb.ObjectTypeScope_PROJECT {
-			cond.ProjectCondition = input.Object.Project
-		}
-	}
-
-	objs, err := s.store.ObjectListWithReferences(
-		[]*conditions.ObjectCondition{cond},
-	)
-	if err != nil {
-		return nil, err
-	}
-	if len(objs) > 1 {
-		return nil, ErrMultipleRecordsFound
-	} else if len(objs) == 1 {
-		existing = objs[0]
-	}
-
-	var changed *types.ObjectWithReferences
-	if existing == nil {
-		s.log.L3(
-			"creating new object of type %s in partition %s with name %s...",
-			input.ObjectType.Code,
-			input.Partition.Uuid,
-			input.Object.Name,
-		)
-		changed, err = s.store.ObjectCreate(input)
-		if err != nil {
-			return nil, err
-		}
-		s.log.L1(
-			"created new object with UUID %s of type %s in partition %s with name %s",
-			changed.Object.Uuid,
-			input.ObjectType.Code,
-			input.Partition.Uuid,
-			input.Object.Name,
-		)
-	} else {
-		if input.Object.Uuid == "" {
-			input.Object.Uuid = existing.Object.Uuid
-		}
-		if objectChanged(existing, input) {
-			s.log.L3("updating object with UUID %s...", existing.Object.Uuid)
-			if changed, err = s.store.ObjectUpdate(input); err != nil {
-				s.log.ERR(
-					"failed updating object with UUID %s: %s",
-					existing.Object.Uuid,
-					err,
-				)
-				return nil, ErrUnknown
-			}
-			s.log.L1(
-				"updated object with UUID %s",
-				changed.Object.Uuid,
-			)
-		} else {
-			changed = existing
-			s.log.L3("no changes to object with UUID %s...", existing.Object.Uuid)
-		}
-	}
-
-	return &pb.ObjectSetResponse{
-		Object: changed.Object,
-	}, nil
-}
-
 // validateObjectCreateRequest ensures that the data the user sent is valid and
 // all referenced projects, partitions, and object types are correct.
 func (s *Server) validateObjectCreateRequest(
@@ -407,19 +161,6 @@ func (s *Server) validateObjectCreateRequest(
 		// We don't want to leak internal implementation errors...
 		s.log.ERR("failed when validating object type in object set: %s", err)
 		return nil, errors.ErrUnknown
-	}
-
-	if obj.Properties != nil {
-		// Validate that the properties set against this object meet any schema
-		// associated with that property key and object type
-		for _, prop := range obj.Properties {
-			_, err := s.validateObjectProperty(
-				part, objType, prop.Key, prop.Value,
-			)
-			if err != nil {
-				return nil, err
-			}
-		}
 	}
 
 	return &types.ObjectWithReferences{
