--- pkg/metadata/types/property_definition.go
@@ -1,156 +1,0 @@
-package types
-
-import (
-	"fmt"
-	"os"
-
-	apitypes "github.com/runmachine-io/runmachine/pkg/api/types"
-	pb "github.com/runmachine-io/runmachine/pkg/metadata/proto"
-)
-
-type PropertyDefinitionMatcher interface {
-	Matches(obj *pb.PropertyDefinition) bool
-}
-
-// PropertyDefinitionWithReferences is a concrete struct containing pointers to
-// already-constructed and validated Partition and ObjectType messages. This is
-// the struct that is passed to backend storage when creating new property
-// schemas, not the protobuffer PropertyDefinition message or the
-// api/types/PropertyDefinition struct, neither of which are guaranteed to be
-// pre-validated and their relations already expanded.
-type PropertyDefinitionWithReferences struct {
-	Partition  *pb.Partition
-	ObjectType *pb.ObjectType
-	Definition *pb.PropertyDefinition
-}
-
-// APItoPBPropertySchema onverts an apitypes PropertySchema to the protobuffer
-// PropertySchema message that will eb stored in backend storage
-func APItoPBPropertySchema(as *apitypes.PropertySchema) *pb.PropertySchema {
-	res := &pb.PropertySchema{
-		Types:   []pb.PropertySchema_Type{},
-		Pattern: as.Pattern,
-	}
-	if len(as.Types) > 0 {
-		for _, astype := range as.Types {
-			switch astype {
-			case "string":
-				res.Types = append(res.Types, pb.PropertySchema_TYPE_STRING)
-			case "integer":
-				res.Types = append(res.Types, pb.PropertySchema_TYPE_INTEGER)
-			case "number":
-				res.Types = append(res.Types, pb.PropertySchema_TYPE_NUMBER)
-			case "boolean":
-				res.Types = append(res.Types, pb.PropertySchema_TYPE_BOOLEAN)
-			default:
-				fmt.Fprintf(
-					os.Stderr,
-					"WARNING: unexpected apitypes PropertySchema type: %s",
-					astype,
-				)
-			}
-		}
-	}
-	if as.MultipleOf != nil {
-		res.MultipleOf = &pb.UInt64Value{
-			Value: uint64(*as.MultipleOf),
-		}
-	}
-	if as.Minimum != nil {
-		res.Minimum = &pb.Int64Value{
-			Value: int64(*as.Minimum),
-		}
-	}
-	if as.Maximum != nil {
-		res.Maximum = &pb.Int64Value{
-			Value: int64(*as.Maximum),
-		}
-	}
-	if as.MinLength != nil {
-		res.MinimumLength = &pb.UInt64Value{
-			Value: uint64(*as.MinLength),
-		}
-	}
-	if as.MaxLength != nil {
-		res.MaximumLength = &pb.UInt64Value{
-			Value: uint64(*as.MaxLength),
-		}
-	}
-	if as.Format != "" {
-		switch as.Format {
-		case "date-time":
-			res.Format = pb.PropertySchema_FORMAT_DATETIME
-		case "date":
-			res.Format = pb.PropertySchema_FORMAT_DATE
-		case "time":
-			res.Format = pb.PropertySchema_FORMAT_TIME
-		case "email":
-			res.Format = pb.PropertySchema_FORMAT_EMAIL
-		case "idn-email":
-			res.Format = pb.PropertySchema_FORMAT_IDN_EMAIL
-		case "hostname":
-			res.Format = pb.PropertySchema_FORMAT_HOSTNAME
-		case "idn-hostname":
-			res.Format = pb.PropertySchema_FORMAT_IDN_HOSTNAME
-		case "ipv4":
-			res.Format = pb.PropertySchema_FORMAT_IPV4
-		case "ipv6":
-			res.Format = pb.PropertySchema_FORMAT_IPV6
-		case "uri":
-			res.Format = pb.PropertySchema_FORMAT_URI
-		case "uri-reference":
-			res.Format = pb.PropertySchema_FORMAT_URI_REFERENCE
-		case "iri":
-			res.Format = pb.PropertySchema_FORMAT_IRI
-		case "iri-reference":
-			res.Format = pb.PropertySchema_FORMAT_IRI_REFERENCE
-		case "uri-template":
-			res.Format = pb.PropertySchema_FORMAT_URI_TEMPLATE
-		default:
-			fmt.Fprintf(
-				os.Stderr,
-				"WARNING: unexpected apitypes PropertySchema format: %s",
-				as.Format,
-			)
-		}
-	}
-	return res
-}
-
-// APItoPBPropertyPermissions converts the apitypes.PropertyPermissions to
-// protobuffer PropertyPermissions that get stored in backend storage
-func APItoPBPropertyPermissions(
-	apiperms []*apitypes.PropertyPermission,
-) []*pb.PropertyPermission {
-	res := make([]*pb.PropertyPermission, len(apiperms))
-	for x, apiperm := range apiperms {
-		// Convert the string "r", "rw" representation to the integer
-		// permission code used in backend protobuffer storage
-		iperm := apitypes.PERMISSION_NONE
-		switch apiperm.Permission {
-		case "r":
-			iperm = apitypes.PERMISSION_READ
-		case "rw":
-			iperm = apitypes.PERMISSION_READ | apitypes.PERMISSION_WRITE
-		case "w":
-			iperm = apitypes.PERMISSION_WRITE
-		default:
-			iperm = apitypes.PERMISSION_NONE
-		}
-		pbperm := &pb.PropertyPermission{
-			Permission: iperm,
-		}
-		if apiperm.Project != "" {
-			pbperm.Project = &pb.StringValue{
-				Value: apiperm.Project,
-			}
-		}
-		if apiperm.Role != "" {
-			pbperm.Role = &pb.StringValue{
-				Value: apiperm.Role,
-			}
-		}
-		res[x] = pbperm
-	}
-	return res
-}
